Facade type hardening, restructure and typed expression
General:
build: added some explicit type casts to deal with compiler warnings, where necessary
chore: corrected some usage of SAL
feat: replaced type erased internal handles of facade types with forward type handles
feat: updated function signatures to consume references, when non-null values are expected and not persistent
feat: reworked the validation function of facade asset types to check for an active internal state and corresponding type yielding
feat: updated constructor signatures of facade assets to propagate the internal asset reference correctly
feat: dropped the managed<T...> type alias of shared_ptr<T...>
refactor: split the inbuild header into separate ones ( one function group or concern per header )
refactor: restructured components to reflect semantic groups and/or inheritance structures
refactor: rename actor component to hierarchy component
refactor: restructured actors and predefined actor classes to generate semantic groups
refactor: restructured assets to reflect semantic groups

+++ src/Engine.Reflect/Compile/HashMap.hpp
@@ -205,7 +205,7 @@ namespace hg {
 				// __noop();
 			}
 
-			return std::distance(start, it);
+			return static_cast<size_t>(std::distance(start, it));
 		}
 
 		[[nodiscard]] constexpr size_t count(const typename key_type::value_type& key_) const noexcept {
